Feat: Validacion de permisos por rol para las notas

+++ BlazorTechNotes.Application/Notes/NoteService.cs
@@ -1,21 +1,53 @@
+using BlazorTechNotes.Application.Exceptions;
+using BlazorTechNotes.Application.Users;
 using BlazorTechNotes.Domain.User;
 
 namespace BlazorTechNotes.Application.Notes;
 
-public class NoteService(INoteRepository noteRepository, IUserRepository userRepository) : INoteService
+public class NoteService(
+  INoteRepository noteRepository, 
+  IUserRepository userRepository,
+  IUserService userService) : INoteService
 {
   private readonly INoteRepository _noteRepository = noteRepository;
   private readonly IUserRepository _userRepository = userRepository;
+  private readonly IUserService _userService = userService;
 
   public async Task<Result<NoteResponse>> CreateNoteAsync(CreateNoteRequest request)
   {
-    Note newNote = request.Adapt<Note>();
-    var note = await _noteRepository.CreateNoteAsync(newNote);
-    return note.Adapt<NoteResponse>();
+    try
+    {
+      Note newNote = request.Adapt<Note>();
+        var userId = await _userService.GetCurrentUserIdAsync();
+
+      if(userId is null)
+        return Result.Fail<NoteResponse>("User not authorized.");
+
+      var isAuthorized = await _userService.CurrentUserCanCreateNoteAsync();
+      if (!isAuthorized)
+        return Result.Fail<NoteResponse>("User not authorized to create notes.");
+
+      newNote.UserId = userId;
+
+      var note = await _noteRepository.CreateNoteAsync(newNote);
+      return note.Adapt<NoteResponse>();
+    }
+    catch (UserNotAuthorizedException)
+    {
+      //TODO: mejorar fail messages
+      return Result.Fail<NoteResponse>("User not authorized.");
+    }
   }
 
   public async Task<Result> DeleteNoteAsync(int id)
   {
+    var userCanDelete = await _userService.CurrentUserCanEditNoteAsync(id);
+
+    if (!userCanDelete)
+    {
+      return Result.Fail<NoteResponse?>("User not authorized to delete this note.");
+    }
+
     var deleted = await _noteRepository.DeleteNoteAsync(id);
 
     if (deleted)
@@ -39,6 +71,8 @@ public class NoteService(INoteRepository noteRepository, IUserRepository userRep
       {
         var author = await _userRepository.GetUserByIdAsync(note.UserId);
         noteResponse.UserName = author?.UserName ?? "Unknown";
+        noteResponse.UserId = note.UserId;
+        noteResponse.CanEdit = await _userService.CurrentUserCanEditNoteAsync(note.Id);
       }
       else
       {
@@ -68,6 +102,8 @@ public class NoteService(INoteRepository noteRepository, IUserRepository userRep
     {
       var user = await _userRepository.GetUserByIdAsync(note.UserId);
       noteResponse.UserName = user?.UserName ?? "Unknown";
+      noteResponse.UserId = note.UserId;
+      noteResponse.CanEdit = await _userService.CurrentUserCanEditNoteAsync(note.Id);
     }
     else
     {
@@ -80,6 +116,13 @@ public class NoteService(INoteRepository noteRepository, IUserRepository userRep
   public async Task<Result<NoteResponse?>> UpdateNoteAsync(UpdateNoteRequest request)
   {
     Note noteToUpdate = request.Adapt<Note>();
+    var userCanEdit = await _userService.CurrentUserCanEditNoteAsync(noteToUpdate.Id);
+
+    if (!userCanEdit)
+    {
+      return Result.Fail<NoteResponse?>("User not authorized to edit this note.");
+    }
+
     var updatedNote = await _noteRepository.UpdateNoteAsync(noteToUpdate);
 
     if (updatedNote is null)
